refactor: 重构注册 udid、更新设备名称
