+++ app/controllers/udid_controller.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class UdidController < ApplicationController
+  include DeviceAttributes
   include Qrcode
 
   before_action :set_device_xml, only: :create
@@ -52,8 +53,10 @@ class UdidController < ApplicationController
     udid = params[:udid]
 
     new_device = apple_key.register_device(udid, name)
-    if new_device.errors
-      flash[:alter] = new_device.errors.messages[:devices][0]
+    unless new_device.valid?
+      logger.debug "Register failed with errors: #{new_device.errors}"
+      error_message = new_device.errors.messages[:devices][0]
+      flash[:alter] = error_message if error_message.present?
       return render :show, status: :unprocessable_entity
     end
 
@@ -65,21 +68,12 @@ class UdidController < ApplicationController
   def install
   end
 
+  # GET /udid/qrcode
   def qrcode
     render qrcode: udid_index_url, **qrcode_options
   end
 
   private
-
-  def set_device_xml
-    body = request.body.read
-
-    p7sign = OpenSSL::PKCS7.new(body)
-    store = OpenSSL::X509::Store.new
-    p7sign.verify(nil, store, nil, OpenSSL::PKCS7::NOVERIFY)
-
-    @device_attrs = Plist.parse_xml(p7sign.data)
-  end
 
   def render_profile
     enable_tls = params[:tls] == 'true'
@@ -97,56 +91,14 @@ class UdidController < ApplicationController
     render plain: sigend.to_der, content_type: content_type, layout: false
   end
 
-  def certificate
-    @certificate ||= lambda do
-      cert = OpenSSL::X509::Certificate.new
-      cert.version = 2
-      cert.serial = Random.rand(2**16 - 2) + 1
-      cert.subject = subject
-      cert.issuer = issuer
-      cert.not_before = create_date
-      cert.not_after = expire_date
-      cert.public_key = root_key.public_key
+  def set_device_xml
+    body = request.body.read
 
-      cert_ext = OpenSSL::X509::ExtensionFactory.new
-      cert_ext.subject_certificate = cert
-      cert_ext.issuer_certificate = cert
-      cert.add_extension(cert_ext.create_extension("basicConstraints","CA:TRUE",true))
-      cert.add_extension(cert_ext.create_extension("keyUsage","keyCertSign, cRLSign", true))
-      cert.add_extension(cert_ext.create_extension("subjectKeyIdentifier","hash",false))
-      cert.add_extension(cert_ext.create_extension("authorityKeyIdentifier","keyid:always",false))
-      cert.sign(root_key, digest)
-      cert
-    end.call
-  end
+    p7sign = OpenSSL::PKCS7.new(body)
+    store = OpenSSL::X509::Store.new
+    p7sign.verify(nil, store, nil, OpenSSL::PKCS7::NOVERIFY)
 
-  def subject
-    @subject ||= OpenSSL::X509::Name.parse("/C=CN/O=tryzealot/OU=zealot/CN=#{Setting.site_title} Self-signed Authority")
-  end
-
-  def issuer
-    @issuer ||= OpenSSL::X509::Name.parse("/C=CN/O=tryzealot/OU=zealot/CN=#{Setting.site_title}")
-  end
-
-  def root_key
-    @key ||= OpenSSL::PKey::RSA.generate(2048)
-  end
-
-  def digest
-    @digest ||= OpenSSL::Digest.new('SHA256')
-  end
-
-  def payload_uuid
-    @payload_uuid ||= Rails.cache.fetch('ios-udid', expires_in: 1.week) { SecureRandom.uuid.upcase }
-  end
-
-  def create_date
-    Time.now.utc
-  end
-
-  def expire_date
-    expire_time = 1 * 365 * 24 * 60 * 60  # 1 year
-    create_date + expire_time
+    @device_attrs = Plist.parse_xml(p7sign.data)
   end
 
   def set_device_metadata
@@ -163,16 +115,13 @@ class UdidController < ApplicationController
   end
 
   def device_status
-    if @apple_keys
-      # :registered_device
-      'related_apple_keys'
-    elsif @all_apple_keys.size > 0
-      # :new_register_device
-      'register_apple_key'
-    else
-      # :new_device
-      'unregister_device'
-    end
+    @device_status = if @apple_keys
+                       :registered_device
+                     elsif @all_apple_keys.size > 0
+                       :new_register_device
+                     else
+                       :new_device
+                     end
   end
 
   def set_apple_key
