+++ app/views/admin/apple_keys/show.html.slim
@@ -65,9 +65,8 @@ ruby:
                 td = device.channels.count
                 td = device.created_at
       .card-footer
-        p
-          - if @apple_key.last_synced_at.present?
-            span title="#{@apple_key.last_synced_at}" data-toggle="tooltip"
-              = t('.last_synced_at', time: time_ago_in_words(@apple_key.last_synced_at))
-          - else
-            = t('.never_sync')
+        - if @apple_key.last_synced_at.present?
+          span title="#{@apple_key.last_synced_at}" data-toggle="tooltip"
+            = t('.last_synced_at', time: time_ago_in_words(@apple_key.last_synced_at))
+        - else
+          = t('.never_sync')
