+++ app/views/udid/_new_device.html.slim
@@ -1,16 +1,12 @@
-- content_for :title do
-  = @title
-
-- content_for :section_title do
-  = @title
-
 .row
+  .col-md-12
+    .alert.alert-warning
+      strong Ta-Da！
+      = t('udid.show.unregistered')
+
   == render 'dashboards/widget', title: t('udid.show.channels'), count: @channel_total, url: '', color_style: 'bg-navy', icon: 'fab fa-adn'
   == render 'dashboards/widget', title: t('udid.show.releases'), count: @release_total, url: '', color_style: 'bg-pink', icon: 'fab fa-gg'
   == render 'dashboards/widget', title: t('udid.show.apple_keys'), count: @apple_key_total, url: '', color_style: 'bg-yellow', icon: 'fas fa-key'
 
-  == render 'edit_form'
-
-  == render 'udid_metadata', device: @device
-  == render 'apple_keys', device: @device, apple_keys: @apple_keys || @all_apple_keys, result: @device_status
-  == render 'apps', device: @device, channels: @channels
+  == render 'udid/udid_metadata', device: @device
+  == render 'udid/apps', device: @device, channels: @channels
