+++ app/views/udid/_udid_metadata.html.slim
@@ -6,20 +6,23 @@
         = t('udid.show.metadata')
     .card-body
       dl.system-info.p-0
-        == render 'item', key: 'udid', value: params[:udid]
+        == render 'udid/item', key: 'udid', value: params[:udid], copy: true
         - if device.present?
           - if device.name
-            == render 'item', key: 'name', value: device.name
+            == render 'udid/item', key: 'name', value: device.name, copy: true
           - if device.platform
-            == render 'item', key: 'platform', value: device.platform
+            == render 'udid/item', key: 'platform', value: device.platform, copy: true
           - if device.model
-            == render 'item', key: 'model', value: device.model
+            == render 'udid/item', key: 'model', value: device.model, copy: true
+          - if device.status
+            == render 'udid/item', key: 'status', value: t("udid.show.#{device.status.downcase}"), copy: false
 
         - if params[:product].present?
-          == render 'item', key: 'product', value: params[:product]
+          == render 'udid/item', key: 'product', value: params[:product], copy: true
         - if params[:version].present?
-          == render 'item', key: 'version', value: params[:version]
+          == render 'udid/item', key: 'version', value: params[:version], copy: true
         - if params[:serial].present?
-          == render 'item', key: 'serial', value: params[:serial]
+          == render 'udid/item', key: 'serial', value: params[:serial], copy: true
+
         - if device.present?
-          == render 'item', key: 'created_at', value: device.created_at
+          == render 'udid/item', key: 'created_at', value: device.created_at, copy: false
