+++ config/locales/zealot/zh-CN.yml
@@ -697,7 +697,7 @@ zh-CN:
           </ol>
     show:
       title: 设备信息
-      metadata: 元数据
+      metadata: 设备信息
       registered_but_no_related_app: 本设备已注册，很遗憾还没有发现关联应用
       registered_with_related_apps:  本设备已注册，下面选择需要安装的应用吧
       register: 本设备还未注册，快去选择对应的苹果开发者主体注册吧！
@@ -710,17 +710,20 @@ zh-CN:
       name: 设备名称
       platform: 平台
       model: 产品型号
+      status: 状态
+      enabled: 可用
+      disabled: 禁用
       created_at: 注册时间
       upload_full_version_release: :'dashboard.timeline.upload_full_version_release'
       upload_auto_version_release: :'dashboard.timeline.upload_auto_version_release'
-      related_apple_keys: 已关联的苹果开发者
-      register_apple_key: 注册测试设备到苹果开发者
+      registered_device: 已关联苹果开发者
+      new_register_device: 注册到苹果开发者
       channels: 应用渠道
       releases: 应用版本
       apple_keys: 苹果开发者
       team_name: 团队名称
       team_id: 团队 ID
-      register_device: 注册本设备
+      register_device: 注册设备
       related_apple_keys_not_found: 没有发现苹果开发者账户
       related_apps_not_found_with_registered_device: 本设备是通过其他设备同步的方式被注册，但您的设备还没有关联的应用上传
       related_apps_not_found: 当前设备可能还没有注册到苹果开发者或注册的应用没有上传到 Zealot
