+++ db/migrate/20230803065502_add_device_id_to_device.rb
@@ -1,0 +1,5 @@
+class AddDeviceIdToDevice < ActiveRecord::Migration[7.0]
+  def change
+    add_column :devices, :device_id, :string
+  end
+end
